Refaktoring. W celu uproszczenia usunieto panel dedykowany dla admina
