--- project/resources/views/admin/warehouses/index.blade.php
@@ -1,79 +1,0 @@
-<x-admin-master>
-
-    @section('custom-styles')
-        <link href="{{asset('vendor/datatables/dataTables.bootstrap4.min.css')}}" rel="stylesheet">    
-    @endsection
-
-    @section('content')
-        <!-- Page Heading -->
-        <h1 class="h3 mb-2 text-gray-800">Lista składów budowlanych</h1>
-        <p class="mb-4">Naciśnij wiersz aby edytować wpis</p>
-
-        <!-- DataTales Example -->
-        <div class="card shadow mb-4">
-          <div class="card-header py-3">
-            <h6 class="m-0 font-weight-bold text-primary">
-                <a href="{{route('warehouses.create')}}">
-                    <span>Nowy</span>
-                    <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-plus-square" viewBox="0 0 16 16">
-                        <path d="M14 1a1 1 0 0 1 1 1v12a1 1 0 0 1-1 1H2a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1h12zM2 0a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V2a2 2 0 0 0-2-2H2z"/>
-                        <path d="M8 4a.5.5 0 0 1 .5.5v3h3a.5.5 0 0 1 0 1h-3v3a.5.5 0 0 1-1 0v-3h-3a.5.5 0 0 1 0-1h3v-3A.5.5 0 0 1 8 4z"/>
-                    </svg>
-                </a>
-            </h6>
-          </div>
-          <div class="card-body">
-            <div class="table-responsive">
-              <table class="table table-bordered table-hover table-sm" id="dataTable" width="100%" cellspacing="0">
-                <thead>
-                  <tr>
-                    <th>Id</th>
-                    <th>Nazwa Sieci</th>
-                    <th>Nazwa składu</th>
-                    <th>Adres</th>
-                    <th>Współrzędne</th>
-                    <th>Stworzono</th>
-                    <th>Zaktualizowano</th>
-                  </tr>
-                </thead>
-                <tfoot>
-                  <tr>
-                    <th>Id</th>
-                    <th>Nazwa Sieci</th>
-                    <th>Nazwa składu</th>
-                    <th>Adres</th>
-                    <th>Współrzędne</th>
-                    <th>Stworzono</th>
-                    <th>Zaktualizowano</th>
-                  </tr>
-                </tfoot>
-                @foreach ($warehouses as $warehouse)
-                <tbody>
-                    <tr onclick="window.location='/admin/warehouses/{{$warehouse->id}}'">
-                        <td>{{$warehouse->id}}</td>
-                        <td>{{$warehouse->company->name}}</td>
-                        <td>{{$warehouse->name}}</td>
-                        <td>{{$warehouse->address}}</td>
-                        <td>{{$warehouse->location}}</td>
-                        <td>{{$warehouse->created_at}}</td>
-                        <td>{{$warehouse->updated_at}}</td>
-                    </tr>
-                  </tbody>    
-                @endforeach
-                
-              </table>
-            </div>
-          </div>
-        </div>
-    @endsection
-
-    @section('custom-scripts')
-        <!-- Page level plugins -->
-        <script src="{{asset('vendor/datatables/jquery.dataTables.min.js')}}""></script>
-        <script src="{{asset('vendor/datatables/dataTables.bootstrap4.min.js')}}"></script>
-
-        <!-- Page level custom scripts -->
-        <script src="{{asset('js/datatables-demo.js')}}"></script>    
-    @endsection
-    
-</x-admin-master>
