+++ project/resources/views/components/addons.blade.php
