--- project/resources/views/components/admin-sidebar-brand.blade.php
@@ -1,10 +1,0 @@
-<!-- Sidebar - Brand -->
-<a class="sidebar-brand d-flex align-items-center justify-content-center" href="{{route('home')}}">
-    <div class="sidebar-brand-icon rotate-n-15">
-      <i class="fas fa-laugh-wink"></i>
-    </div>
-    <div class="sidebar-brand-text mx-3">SB Admin <sup>2</sup></div>
-  </a>
-  
-  <!-- Divider -->
-  <hr class="sidebar-divider my-0">
