+++ project/resources/views/components/components.blade.php
@@ -1,3 +1,6 @@
+<!-- Divider -->
+<hr class="sidebar-divider">
+
 <!-- Heading -->
 <div class="sidebar-heading">
 Baza danych
@@ -6,7 +9,7 @@ Baza danych
 <!-- Nav Item - sieci sklepów -->
 <li class="nav-item">
     <a class="nav-link" href="{{route('companies.index')}}">
-      <i class="fas fa-fw fa-cog"></i>
+      <i class="fas fa-tags"></i>
       <span>Sieci sklepów</span>
     </a>
 </li>
@@ -14,7 +17,7 @@ Baza danych
 <!-- Nav Item - składy budowlane -->
 <li class="nav-item">
   <a class="nav-link" href="{{route('warehouses.index')}}">
-    <i class="fas fa-fw fa-wrench"></i>
+    <i class="fas fa-pallet"></i>
     <span>Składy budowlane</span>
   </a>
 </li>
