+++ project/resources/views/components/dashboard.blade.php
