--- project/resources/views/components/home-categories.blade.php
@@ -1,34 +1,0 @@
-<!-- Categories Widget -->
-<div class="card my-4">
-    <h5 class="card-header">Categories</h5>
-    <div class="card-body">
-      <div class="row">
-        <div class="col-lg-6">
-          <ul class="list-unstyled mb-0">
-            <li>
-              <a href="#">Web Design</a>
-            </li>
-            <li>
-              <a href="#">HTML</a>
-            </li>
-            <li>
-              <a href="#">Freebies</a>
-            </li>
-          </ul>
-        </div>
-        <div class="col-lg-6">
-          <ul class="list-unstyled mb-0">
-            <li>
-              <a href="#">JavaScript</a>
-            </li>
-            <li>
-              <a href="#">CSS</a>
-            </li>
-            <li>
-              <a href="#">Tutorials</a>
-            </li>
-          </ul>
-        </div>
-      </div>
-    </div>
-</div>
