--- project/resources/views/components/home-search.blade.php
@@ -1,12 +1,0 @@
-<!-- Search Widget -->
-<div class="card my-4">
-    <h5 class="card-header">Search</h5>
-    <div class="card-body">
-    <div class="input-group">
-        <input type="text" class="form-control" placeholder="Search for...">
-        <span class="input-group-btn">
-        <button class="btn btn-secondary" type="button">Go!</button>
-        </span>
-    </div>
-    </div>
-</div>
