--- project/resources/views/components/home-side-widget.blade.php
@@ -1,7 +1,0 @@
-<!-- Side Widget -->
-<div class="card my-4">
-    <h5 class="card-header">Side Widget</h5>
-    <div class="card-body">
-      You can put anything you want inside of these side widgets. They are easy to use, and feature the new Bootstrap 4 card containers!
-    </div>
-  </div>
