+++ project/resources/views/components/logout-modal.blade.php
@@ -11,7 +11,10 @@
         <div class="modal-body">Select "Logout" below if you are ready to end your current session.</div>
         <div class="modal-footer">
           <button class="btn btn-secondary" type="button" data-dismiss="modal">Cancel</button>
-          <a class="btn btn-primary" href="/logout">Logout</a>
+          <form action="/logout" method="POST">
+            {{ csrf_field() }}
+            <button type="submit" class="btn btn-primary" href="/logout">Logout</button>
+          </form>
         </div>
       </div>
     </div>
