+++ project/resources/views/components/sidebar-brand.blade.php
@@ -1,0 +1,7 @@
+<!-- Sidebar - Brand -->
+<a class="sidebar-brand d-flex align-items-center justify-content-center" href="{{route('home')}}">
+    <div class="sidebar-brand-icon">
+      <i class="fas fa-warehouse"></i>
+    </div>
+    <div class="sidebar-brand-text mx-3">Składy Budowlane</div>
+</a>
